+++ luadatapacktypes/src/storage.d.ts
@@ -5,7 +5,7 @@ interface Storage {
      * Saves an object to the world storage.
      * 
      * @remarks
-     * Use namespaced ids like `foo:bar/baz`
+     * Use namespaced ids like `foo:bar/baz`, or any other string
      * 
      * @param path Namespaced ID
      * @param obj Object
@@ -16,7 +16,7 @@ interface Storage {
      * Gets an object in world storage.
      * 
      * @remarks
-     * Use namespaced ids like `foo:bar/baz`
+     * Use namespaced ids like `foo:bar/baz`, or any other string
      * 
      * @param path Namespaced ID
      */